bg(api): change folder structure
- changed modules in controllers to be only in two modules

[Continues #168864344]
